requirement.txt
SpeechRecognition
deep-translator
keyboard
spacy
streamlit
watchdog
opencv-python
pillow
pyaudio
